# backend/swagger.yaml
openapi: 3.0.3
info:
  title: API del Restaurante
  version: "1.0.0"
servers:
  - url: "http://localhost:3000"
paths:
  /admin:
    get:
      summary: Listar administradores
      responses:
        '200':
          description: OK
    post:
      summary: Crear administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Creado
  /admin/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener administrador por ID
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar administrador
      responses:
        '200':
          description: Eliminado
...
  /menu:
    get:
      summary: Listar menús
      responses:
        '200':
          description: OK
    post:
      summary: Crear menú
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                menuname:
                  type: string
      responses:
        '201':
          description: Creado
  /menu/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener menú
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar menú
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                menuname:
                  type: string
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar menú
      responses:
        '200':
          description: Eliminado
  /menuitem:
    get:
      summary: Listar ítems de menú
      responses:
        '200':
          description: OK
    post:
      summary: Crear ítem de menú
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                menuid:
                  type: integer
                itemname:
                  type: string
                price:
                  type: number
      responses:
        '201':
          description: Creado
  /menuitem/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener ítem de menú
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar ítem de menú
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                menuid:
                  type: integer
                itemname:
                  type: string
                price:
                  type: number
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar ítem de menú
      responses:
        '200':
          description: Eliminado
...